Backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
openai-whisper==20240930
moviepy==1.0.3